libamelia - Added QDoubleSync. Works like QBoolSync, but with doubles.
ageometry
- Possible to edit a combination name through the table.
- We no longer ask for a combination name upon creation
- Hide detector after loading an event. Temporary solution until wireframe comes along.
- Use QDoubleSync to synchronize the pt cut input devices.

aeventmanager
- ATrackCombination now returns a default name as "track1|track2|track3"
- APtFilter now uses doubles instead of floats
- Added AEventAnalysisData::forceUpdate() to force saving of the package.

+++ trunk/src/aeventmanager/APtFilter.h
@@ -11,15 +11,14 @@ class AEVENTMANAGER_EXPORT APtFilter : public ATrackFilter
   int _minPt;
   
  public:
-  APtFilter(float minPt,ATrackFilter* next=0);
+  APtFilter(double minPt,ATrackFilter* next=0);
 
-  float minPt();
+  double minPt();
   bool checkTrack(ATrack* track);
   
  public slots:
-  void setMinPt(float);
-  void setMinPtMeV(int);
-  
+  void setMinPt(double);
+  void setMinPtMeV(double);
 };
 
 #endif //APTFILTER_H_
